refactor(bot): extract ChargedHoldReleaseCast, retrofit Fireball/Zoltraak
